requirements.txt
pyqt >=5.15
pyqtwebengine >=5.15
pyinstaller >=6
pymolpro >=1.17.9
pysjef >=1.44.0
pubchempy >=1
chemspipy >=2
git >= 2.49
libboost >= 1.84.0
certifi >= 2024.8.30
jsonschema >= 4
openssl >= 3.5.3
